requirement.txt
# utility 
tqdm==4.66.3
numpy==1.24.1
natsort==8.4.0
lightning==2.4.0
tensorboard==2.18.0

#image processing
scikit-image==0.21.0
imageio==2.31.1
Pillow==10.2.0
pyshtools==4.13.1

# EXR handling 
skylibs==0.7.4
OpenEXR==1.3.9

# Diffusers dependencies
accelerate==1.0.1
datasets==3.0.1
diffusers==0.30.3
transformers==4.45.2
xformers==0.0.28.post1
bitsandbytes==0.44.0
